# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is about to project management",
    "version": "1.0.0",
    "title": "Project Management",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "calm-brook-82945.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/projects/addproject": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "add new project",
        "description": "",
        "operationId": "add new project",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "adding new project schema",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addproject"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/projects/deleteproject": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "delete project",
        "description": "",
        "operationId": "delete project",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "delete project",
            "required": true,
            "schema": {
              "$ref": "#/definitions/projectdelete"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "addproject": {
      "type": "object",
      "properties": {
        "appname": {
          "type": "string"
        },
        "client": {
          "type": "string"
        },
        "timelimit": {
          "type": "string"
        },
        "startdate": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "xml": {
        "name": "addproject"
      }
    },
...
    "projectdelete": {
      "type": "object",
      "properties": {
        "appname": {
          "type": "string"
        },
        "client": {
          "type": "string"
        }
      },
      "xml": {
        "name": "projectdelete"
      }
    },
...
  }
}